pyproject.toml
[project]
name = "openfl2.0"
version = "0.1.0"
description = "OpenFL 2.0"
readme = "README.md"
requires-python = "==3.10.*"
license = { text = "MIT" }

# requirements
dependencies = [
  "torch==2.9.0",
  "torchvision==0.24.0",
  "numpy==2.2.6",
  "web3==7.14.0",
  "termcolor==3.1.0",
  "matplotlib==3.10.7",
  "eth-abi==5.2.0",
  "pandas==2.3.3",
  "py-solc-x==2.0.4",
  "python-dotenv==1.2.1"
]

# Dev-only tools (notebooks, testing, formatting, etc.)
[project.optional-dependencies]
dev = [
  "jupyter==1.1.1",
  "jupyterlab==4.4.10",
  "ipykernel",
  "black",
  "ruff",
  "mypy",
  "pytest",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

